Extract common protocol methods into own class
