+++ j5/backends/hardware/sr/v4/serial/__init__.py
@@ -1,3 +1,5 @@
+"""Serial Backends for Student Robotics version 4 boards in the hardware environment."""
+
 from .power_board import SRV4SerialProtocolPowerBoardHardwareBackend
 
 __all__ = ["SRV4SerialProtocolPowerBoardHardwareBackend"]
